[issue-39] reuse httpclient between invocations
This change makes the Slack object reuse it's SlackHttpClient.
Notice that even if Slack object is provided a SlackHttpClient in its builder function,
this instance was never used.

Recreating the underlying OkHttpClient is a very expensive operation,
setting up threadpools and re-establishing TCP connections/https handshakes,
and should be avoided.