Add access to `SlackApiLambdaHandler` internal `App`
Adds a protected method `SlackApiLambdaHandler#app` to allow lambda
handlers implementing that class to modify the behaviour of the
underlying `App` after initialisation.

Moves the call to `App#start` to the constructor to avoid unnecessarily
call it every time a warmup event triggers the lambda. It is only needed
to run it once when the lambda starts up the first time.

Reorders the methods in `SlackApiLambdaHandler` to be next to the
calling method, enhacing readability.

Resolves: https://github.com/slackapi/java-slack-sdk/issues/419