Fix race conditions by adding atomicity
While a ConcurrentHashMap is a hash table that supports full concurrency
of retrievals and updates where .get() and .put() themselves are both
thread-safe, the particular order of operations being executed in these
methods are not being performed atomically.

It's important that we perform these operations atomically by using
atomic methods from the ConcurrentHashMap class.

Most of the logic in these methods can be replaced by one-line atomic
operations that eliminate very difficult to replicate race conditions
created from the lack of atomicity.